se mueve el elemnto de la card individual a su propio archivo
